# openapi.json
{
    "components": {
        "schemas": {
...
            "ErrorDetail": {
                "additionalProperties": false,
                "properties": {
                    "location": {
                        "description": "Where the error occurred, e.g. 'body.items[3].tags' or 'path.thing-id'",
                        "type": "string"
                    },
                    "message": {
                        "description": "Error message text",
                        "type": "string"
                    },
                    "value": {
                        "description": "The value at the given location"
                    }
                },
                "type": "object"
            },
            "ErrorModel": {
                "additionalProperties": false,
                "properties": {
                    "$schema": {
                        "description": "A URL to the JSON Schema for this object.",
                        "examples": [
                            "https://example.com/schemas/ErrorModel.json"
                        ],
                        "format": "uri",
                        "readOnly": true,
                        "type": "string"
                    },
                    "detail": {
                        "description": "A human-readable explanation specific to this occurrence of the problem.",
                        "examples": [
                            "Property foo is required but is missing."
                        ],
                        "type": "string"
                    },
                    "errors": {
                        "description": "Optional list of individual error details",
                        "items": {
                            "$ref": "#/components/schemas/ErrorDetail"
                        },
                        "type": [
                            "array",
                            "null"
                        ]
                    },
                    "instance": {
                        "description": "A URI reference that identifies the specific occurrence of the problem.",
                        "examples": [
                            "https://example.com/error-log/abc123"
                        ],
                        "format": "uri",
                        "type": "string"
                    },
                    "status": {
                        "description": "HTTP status code",
                        "examples": [
                            400
                        ],
                        "format": "int64",
                        "type": "integer"
                    },
                    "title": {
                        "description": "A short, human-readable summary of the problem type. This value should not change between occurrences of the error.",
                        "examples": [
                            "Bad Request"
                        ],
                        "type": "string"
                    },
                    "type": {
                        "default": "about:blank",
                        "description": "A URI reference to human-readable documentation for the error.",
                        "examples": [
                            "https://example.com/errors/example"
                        ],
                        "format": "uri",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "FileOut": {
                "additionalProperties": false,
                "properties": {
                    "dir": {
                        "description": "True when the entry is a directory",
                        "type": "boolean"
                    },
                    "path": {
                        "description": "Path relative to the project root",
                        "type": "string"
                    },
                    "size": {
                        "description": "Size in bytes",
                        "format": "int64",
                        "type": "integer"
                    }
                },
                "required": [
                    "path",
                    "size",
                    "dir"
                ],
                "type": "object"
            },
...
            "ListFilesOutputBody": {
                "additionalProperties": false,
                "properties": {
                    "$schema": {
                        "description": "A URL to the JSON Schema for this object.",
                        "examples": [
                            "https://example.com/schemas/ListFilesOutputBody.json"
                        ],
                        "format": "uri",
                        "readOnly": true,
                        "type": "string"
                    },
                    "items": {
                        "items": {
                            "$ref": "#/components/schemas/FileOut"
                        },
                        "type": [
                            "array",
                            "null"
                        ]
                    }
                },
                "required": [
                    "items"
                ],
                "type": "object"
            },
            "ListProjectsOutputBody": {
                "additionalProperties": false,
                "properties": {
                    "$schema": {
                        "description": "A URL to the JSON Schema for this object.",
                        "examples": [
                            "https://example.com/schemas/ListProjectsOutputBody.json"
                        ],
                        "format": "uri",
                        "readOnly": true,
                        "type": "string"
                    },
                    "items": {
                        "items": {
                            "$ref": "#/components/schemas/ProjectOut"
                        },
                        "type": [
                            "array",
                            "null"
                        ]
                    }
                },
                "required": [
                    "items"
                ],
                "type": "object"
            },
            "ProjectOut": {
                "additionalProperties": false,
                "properties": {
                    "created_at": {
                        "description": "Last modification time of the project directory",
                        "format": "date-time",
                        "type": "string"
                    },
                    "language": {
                        "description": "Detected language hint (latex / go / cpp / python / rust / javascript / markdown)",
                        "type": "string"
                    },
                    "name": {
                        "description": "Project name (filesystem-safe identifier)",
                        "type": "string"
                    }
                },
                "required": [
                    "name",
                    "created_at"
                ],
                "type": "object"
            },
...
            "StartCompileBody": {
                "additionalProperties": false,
                "properties": {
                    "$schema": {
                        "description": "A URL to the JSON Schema for this object.",
                        "examples": [
                            "https://example.com/schemas/StartCompileBody.json"
                        ],
                        "format": "uri",
                        "readOnly": true,
                        "type": "string"
                    },
                    "bib": {
                        "description": "Bibliography processor : bibtex | biber. Defaults to bibtex.",
                        "type": "string"
                    },
                    "command": {
                        "description": "Optional verbatim shell command \u2014 overrides the language's default when set. Inside the workspace \u03bcVM, the dispatcher runs sh -c <command>.",
                        "type": "string"
                    },
                    "engine": {
                        "description": "LaTeX engine to invoke for the latex language : pdflatex | lualatex | xelatex. Defaults to pdflatex.",
                        "type": "string"
                    },
                    "entry": {
                        "description": "Main source file (e.g. main.tex, main.go). Per-language default applies when empty.",
                        "type": "string"
                    },
                    "extra_args": {
                        "description": "Appended to the language's default build command.",
                        "items": {
                            "type": "string"
                        },
                        "type": [
                            "array",
                            "null"
                        ]
                    },
                    "language": {
                        "description": "Sandbox image picker (latex / go / cpp / python / rust / javascript). Required.",
                        "type": "string"
                    }
                },
                "required": [
                    "language"
                ],
                "type": "object"
            },
            "StartCompileOutputBody": {
                "additionalProperties": false,
                "properties": {
                    "$schema": {
                        "description": "A URL to the JSON Schema for this object.",
                        "examples": [
                            "https://example.com/schemas/StartCompileOutputBody.json"
                        ],
                        "format": "uri",
                        "readOnly": true,
                        "type": "string"
                    },
                    "id": {
                        "description": "Job identifier ; tail logs via GET /api/projects/{name}/compile/{id} (SSE).",
                        "type": "string"
                    }
                },
                "required": [
                    "id"
                ],
                "type": "object"
            },
...
        }
    },
    "info": {
        "title": "weft-loom-server API",
        "version": "v1"
    },
    "openapi": "3.1.0",
    "paths": {
...
        "/api/projects": {
            "get": {
                "operationId": "list-projects",
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ListProjectsOutputBody"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "default": {
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorModel"
                                }
                            }
                        },
                        "description": "Error"
                    }
                },
                "summary": "List the caller's projects",
                "tags": [
                    "projects"
                ]
            }
        },
...
        "/api/projects/{name}/compile": {
            "post": {
                "description": "Schedules an ephemeral compile in the language's sandbox image. Returns the job ID immediately ; logs + result stream over Server-Sent Events at /api/projects/{name}/compile/{id} (raw, outside the typed API).",
                "operationId": "start-compile",
                "parameters": [
                    {
                        "description": "Project name",
                        "in": "path",
                        "name": "name",
                        "required": true,
                        "schema": {
                            "description": "Project name",
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StartCompileBody"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "202": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StartCompileOutputBody"
                                }
                            }
                        },
                        "description": "Accepted"
                    },
                    "default": {
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorModel"
                                }
                            }
                        },
                        "description": "Error"
                    }
                },
                "summary": "Start a compile job",
                "tags": [
                    "compile"
                ]
            }
        },
        "/api/projects/{name}/files": {
            "get": {
                "description": "Returns every file in the project as a flat list with relative paths. Directories appear with dir=true. Binary read/write of one file lands on /api/projects/{name}/files/{path...} (raw, outside the typed API).",
                "operationId": "list-files",
                "parameters": [
                    {
                        "description": "Project name",
                        "in": "path",
                        "name": "name",
                        "required": true,
                        "schema": {
                            "description": "Project name",
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ListFilesOutputBody"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "default": {
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorModel"
                                }
                            }
                        },
                        "description": "Error"
                    }
                },
                "summary": "List files in a project",
                "tags": [
                    "files"
                ]
            }
        },
...
    }
}